requirements.txt
fastapi==0.78.0
pydantic>=1.8.0,<2.0.0
uvicorn>=0.15.0,<0.16.0
Pillow==9.1.1
numpy==1.23.0
requests==2.28.0
jinja2==3.1.2
